CMS-577: nav guard, confirmation, flash on non-winter
