+++ frontend/src/router/pages/SeasonPage.jsx
@@ -1,11 +1,17 @@
 import { useEffect, useState } from "react";
 import { Outlet, useNavigate, useParams } from "react-router-dom";
+import omit from "lodash/omit";
 
 import paths from "@/router/paths";
-import { useApiGet } from "@/hooks/useApi";
+import { useApiGet, useApiPost } from "@/hooks/useApi";
 import useValidation from "@/hooks/useValidation";
+import { useFlashMessage } from "@/hooks/useFlashMessage";
+import { useConfirmation } from "@/hooks/useConfirmation";
+import { useNavigationGuard } from "@/hooks/useNavigationGuard";
 
 import LoadingBar from "@/components/LoadingBar";
+import FlashMessage from "@/components/FlashMessage";
+import ConfirmationDialog from "@/components/ConfirmationDialog";
 
 export default function SeasonPage() {
   // Render the SubmitDates (edit form) or PreviewDates (read-only) component
@@ -14,15 +20,25 @@ export default function SeasonPage() {
 
   const { parkId, seasonId } = useParams();
   const navigate = useNavigate();
+  const errorFlashMessage = useFlashMessage();
+  const confirmationDialog = useConfirmation();
 
   const [season, setSeason] = useState(null);
   const [dates, setDates] = useState(null);
   const [readyToPublish, setReadyToPublish] = useState(false);
   const [notes, setNotes] = useState("");
 
+  // Track deleted date ranges
+  const [deletedDateRangeIds, setDeletedDateRangeIds] = useState([]);
+
   const validation = useValidation(dates, notes, season);
 
   const { data, loading, error } = useApiGet(`/seasons/${seasonId}`);
+
+  const { sendData, loading: saving } = useApiPost(
+    // @TODO: refactor too
+    `/seasons/${seasonId}/save/`,
+  );
 
   useEffect(() => {
     if (data) {
@@ -76,10 +92,55 @@ export default function SeasonPage() {
     window.scrollTo(0, 0);
   }
 
-  async function saveAsDraft(saveFunction, openConfirmation, showErrorFlash) {
+  async function saveChanges() {
+    // Build a list of date ranges of all date types
+    const allDates = Object.values(dates)
+      .reduce(
+        (acc, dateType) => acc.concat(dateType.Operation, dateType.Reservation),
+        [],
+      )
+      // Filter out any blank ranges or unchanged dates
+      .filter((dateRange) => {
+        if (
+          dateRange.startDate === null &&
+          dateRange.endDate === null &&
+          !dateRange.id
+        ) {
+          return false;
+        }
+
+        // if the range is unchanged, skip this range
+        return dateRange.changed;
+      })
+      // Add dateTypeId to the date ranges, remove tempId
+      .map((dateRange) => ({
+        ...omit(dateRange, ["tempId"]),
+        dateTypeId: dateRange.dateType.id,
+      }));
+
+    const payload = {
+      notes,
+      readyToPublish,
+      dates: allDates,
+      deletedDateRangeIds,
+    };
+
+    const response = await sendData(payload);
+
+    return response;
+  }
+
+  function showErrorFlash() {
+    errorFlashMessage.openFlashMessage(
+      "Unable to save changes",
+      "There was a problem saving these changes. Please try again.",
+    );
+  }
+
+  async function saveAsDraft() {
     try {
       if (["pending review", "approved", "on API"].includes(season.status)) {
-        const confirm = await openConfirmation(
+        const confirm = await confirmationDialog.openConfirmation(
           "Move back to draft?",
           "The dates will be moved back to draft and need to be submitted again to be reviewed.",
           "Move to draft",
@@ -90,7 +151,7 @@ export default function SeasonPage() {
         if (!confirm) return;
       }
 
-      await saveFunction();
+      await saveChanges();
 
       navigate(`${paths.park(parkId)}?saved=${seasonId}`);
     } catch (err) {
@@ -105,6 +166,33 @@ export default function SeasonPage() {
       }
     }
   }
+
+  // Returns true if there are any form changes to save
+  function hasChanges() {
+    if (!dates) return false;
+
+    // Any existing dates changed
+    if (
+      Object.values(dates).some((dateType) =>
+        dateType.Operation.concat(dateType.Reservation).some(
+          (dateRange) => dateRange.changed,
+        ),
+      )
+    ) {
+      return true;
+    }
+
+    // If any date ranges have been deleted
+    if (deletedDateRangeIds.length > 0) return true;
+
+    // Ready to publish state has changed
+    if (readyToPublish !== season.readyToPublish) return true;
+
+    // Notes have been entered
+    return notes;
+  }
+
+  useNavigationGuard(hasChanges, confirmationDialog.openConfirmation);
 
   if (loading || !season) {
     return (
@@ -123,22 +211,47 @@ export default function SeasonPage() {
   }
 
   return (
-    <Outlet
-      context={{
-        parkId,
-        seasonId,
-        season,
-        dates,
-        setDates,
-        notes,
-        setNotes,
-        readyToPublish,
-        setReadyToPublish,
-        validation,
-        navigate,
-        navigateAndScroll,
-        saveAsDraft,
-      }}
-    />
+    <>
+      <FlashMessage
+        title={errorFlashMessage.flashTitle}
+        message={errorFlashMessage.flashMessage}
+        isVisible={errorFlashMessage.isFlashOpen}
+        onClose={errorFlashMessage.handleFlashClose}
+        variant="error"
+      />
+
+      <ConfirmationDialog
+        title={confirmationDialog.title}
+        message={confirmationDialog.message}
+        confirmButtonText={confirmationDialog.confirmButtonText}
+        cancelButtonText={confirmationDialog.cancelButtonText}
+        notes={confirmationDialog.confirmationDialogNotes}
+        onCancel={confirmationDialog.handleCancel}
+        onConfirm={confirmationDialog.handleConfirm}
+        isOpen={confirmationDialog.isConfirmationOpen}
+      />
+
+      <Outlet
+        context={{
+          parkId,
+          seasonId,
+          season,
+          dates,
+          setDates,
+          notes,
+          setNotes,
+          readyToPublish,
+          setReadyToPublish,
+          validation,
+          navigate,
+          navigateAndScroll,
+          setDeletedDateRangeIds,
+          saveAsDraft,
+          showErrorFlash,
+          hasChanges,
+          saving,
+        }}
+      />
+    </>
   );
 }
